enums.py  __init__.py  programme.py  student.py
So first of all,  i updated the student.py by including gender and creating an enum.py to contain a predefined data which will be used. 
Also, i took to myself of helping @360konet with the programme model when he was getting errors. Even though he was able to create it, i think to be safe some data columns were just made as str.